% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{add_metadata}
\alias{add_metadata}
\title{Add metadata to the cevodata object}
\usage{
add_metadata(object, data)
}
\arguments{
\item{object}{object}

\item{data}{name of new default assay}
}
\description{
Add metadata to the cevodata object
}
