% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevodata-class.R
\name{adding_components}
\alias{adding_components}
\alias{add_SNV_data}
\alias{add_CNA_data}
\alias{add_models}
\title{Get/Add SNV/CNA data from the cevodata dataset}
\usage{
add_SNV_data(object, snvs, name = NULL)

add_CNA_data(object, cnas, name = NULL)

add_models(object, models, name)
}
\arguments{
\item{object}{object}

\item{snvs}{tibble with SNVs}

\item{name}{name for SNVs/CNAs assay}

\item{cnas}{tibble with CNAs}

\item{models}{cv_subitem object with models}

\item{...}{other arguments}
}
\description{
Get/Add SNV/CNA data from the cevodata dataset
}
\section{Functions}{
\itemize{
\item \code{add_SNV_data()}: Add new SNVs to cevodata

\item \code{add_CNA_data()}: Add new CNAs to cevodata

\item \code{add_models()}: Add new models to cevodata

}}
