% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvs.R, R/getters.R, R/snvs.R
\name{add_CNA_data}
\alias{add_CNA_data}
\alias{SNVs}
\alias{SNVs.cevodata}
\alias{get_SNVs_wider}
\alias{CNAs}
\alias{CNAs.cevodata}
\alias{assays}
\alias{add_SNV_data}
\title{Get/Add SNV/CNA data from the cevodata dataset}
\usage{
add_CNA_data(object, cnas, name = NULL)

SNVs(object, ...)

\method{SNVs}{cevodata}(object, which = default_SNVs(object), ...)

get_SNVs_wider(object, fill_na = NULL)

CNAs(object, ...)

\method{CNAs}{cevodata}(object, which = default_CNAs(object), ...)

add_SNV_data(object, snvs, name = NULL)
}
\arguments{
\item{object}{object}

\item{cnas}{tibble with CNAs}

\item{name}{name for SNVs/CNAs assay}

\item{which}{assay to use - uses active_SNVs if none}

\item{fill_na}{fill missing with this value}

\item{snvs}{tibble with SNVs}
}
\value{
tibble
}
\description{
Get/Add SNV/CNA data from the cevodata dataset
}
\section{Functions}{
\itemize{
\item \code{add_CNA_data()}: Add new CNAs to cevodata

\item \code{SNVs(cevodata)}: Get SNVs from cevodata dataset

\item \code{get_SNVs_wider()}: Get SNVs in the wide table form

\item \code{CNAs(cevodata)}: Get CNAs from cevodata dataset

\item \code{add_SNV_data()}: Add new SNVs to cevodata

}}
