% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevodata-class.R
\name{cevo_metadata}
\alias{cevo_metadata}
\alias{add_patient_data}
\alias{add_sample_data}
\title{Add metadata to the cevodata object}
\usage{
add_patient_data(object, data)

add_sample_data(object, data)
}
\arguments{
\item{object}{object}

\item{data}{name of new default assay}
}
\description{
Add metadata to the cevodata object
}
\section{Functions}{
\itemize{
\item \code{add_patient_data()}: Add patient data to cevodata object

\item \code{add_sample_data()}: Add samples' data to cevodata object

}}
