% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevodata-package.R
\docType{package}
\name{cevodata-package}
\alias{cevodata}
\alias{cevodata-package}
\title{cevodata: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Paweł Kuś \email{kpawel2210@gmail.com}

}
\keyword{internal}
