% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevodata-class.R
\name{data_assays}
\alias{data_assays}
\alias{add_SNV_data}
\alias{add_CNA_data}
\title{Get/Add SNV/CNA data from the cevodata dataset}
\usage{
add_SNV_data(object, snvs, name = NULL)

add_CNA_data(object, cnas, name = NULL)
}
\arguments{
\item{object}{object}

\item{snvs}{tibble with SNVs}

\item{name}{name for SNVs/CNAs assay}

\item{cnas}{tibble with CNAs}

\item{which}{assay to use - uses active_SNVs if none}
}
\description{
Get/Add SNV/CNA data from the cevodata dataset
}
\section{Functions}{
\itemize{
\item \code{add_SNV_data()}: Add new SNVs to cevodata

\item \code{add_CNA_data()}: Add new CNAs to cevodata

}}
