% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevodata-exports.R
\name{export_to_clip}
\alias{export_to_clip}
\title{Export cevodata to CliP input}
\usage{
export_to_clip(
  cd,
  out_dir = NULL,
  snvs_name = default_SNVs(cd),
  cnas_name = default_CNAs(cd),
  purity_column = "purity",
  keep_chromosomes = str_c("chr", 1:22)
)
}
\arguments{
\item{cd}{cevodata object}

\item{out_dir}{Directory where to save files. List of items is returned if
out_dir is NULL}

\item{snvs_name}{name of the snvs to use}

\item{cnas_name}{name of the cnas to use}

\item{purity_column}{name of the metadata column with the purity estimates
to be used}

\item{keep_chromosomes}{list of non-sex chromosomes. CliP does not use sex
chromosomes}
}
\description{
\href{https://github.com/wwylab/CliP}{CliP} is an algorithm for clonal structure
identification through penalizing pairwise differences by Wenyi Wang Lab
at MD Anderson Cancer Center in Houston.
}
