% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvs.R
\name{get_CNAs_var_names}
\alias{get_CNAs_var_names}
\title{Get CNA variable names from cevodata object}
\usage{
get_CNAs_var_names(object, which = default_CNAs(object), ...)
}
\arguments{
\item{which}{CNA assay to use}
}
\description{
Get CNA variable names from cevodata object
}
