% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_patients_data}
\alias{get_patients_data}
\title{Get metadata cols that are constant within each patient}
\usage{
get_patients_data(cd)
}
\arguments{
\item{cd}{\if{html}{\out{<cevodata>}} object}
}
\description{
Get metadata cols that are constant within each patient
}
