% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevodata-class.R
\name{init_cevodata}
\alias{init_cevodata}
\title{Create new cevomod dataset object}
\usage{
init_cevodata(
  name = "Unnamed dataset",
  snvs = NULL,
  snvs_name = NULL,
  cnas = NULL,
  cnas_name = NULL
)
}
\arguments{
\item{name}{dataset name}

\item{snvs}{tibble with SNVs}

\item{snvs_name}{name for SNVs assay}

\item{cnas}{tibble with CNAs}

\item{cnas_name}{name for CNAs assay}
}
\value{
\code{cevodata} object
}
\description{
Create new cevomod dataset object
}
