% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{join_metadata}
\alias{join_metadata}
\title{Join object metadata to a tibble}
\usage{
join_metadata(tbl, cd, by = "sample_id")
}
\arguments{
\item{tbl}{Tibble}

\item{cd}{\if{html}{\out{<cevodata>}} object}

\item{by}{Column name to join by}
}
\description{
Join object metadata to a tibble
}
