% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_williams_Mf_1f.R
\name{Mf_1f}
\alias{Mf_1f}
\alias{calc_Mf_1f}
\alias{calc_Mf_1f.cevodata}
\alias{calc_Mf_1f.cevo_snvs}
\alias{get_Mf_1f}
\alias{plot_Mf_1f}
\alias{plot_Mf_1f.cevodata}
\title{William's M(f) ~ 1/f statistics}
\usage{
calc_Mf_1f(object, ...)

\method{calc_Mf_1f}{cevodata}(
  object,
  name = "Mf_1f",
  snvs_name = default_SNVs(object),
  column = get_snvs_frequency_measure(object, snvs_name),
  bins = 100,
  verbose = get_verbosity(),
  ...
)

\method{calc_Mf_1f}{cevo_snvs}(
  object,
  column = get_snvs_frequency_measure(object),
  bins = 100,
  verbose = get_verbosity(),
  ...
)

get_Mf_1f(object, name = "Mf_1f", verbose = get_verbosity(), ...)

plot_Mf_1f(object, ...)

\method{plot_Mf_1f}{cevodata}(
  object,
  bins = NULL,
  from = 0.1,
  to = 0.25,
  scale = TRUE,
  geom = "point",
  ...
)
}
\arguments{
\item{object}{SNVs tibble object}

\item{...}{Other arguments}

\item{name}{name of slot with SFS statistics}

\item{snvs_name}{Which SNVs to use?}

\item{column}{VAF or CCF/2}

\item{bins}{Resolution of the cumulative tails calculation}

\item{verbose}{verbose?}

\item{from}{min f to plot}

\item{to}{max f to plot}

\item{scale}{scale data?}

\item{geom}{ggplot geom to use, eg. geom_line()}
}
\value{
ggplot obj
}
\description{
Mf_1f columns description:
\itemize{
\item sample_id
\item f
\item n - number of mutations in the VAF interval
\item \code{M(f)} and \code{1/f} columns to plot William's statistics
}
}
\section{Functions}{
\itemize{
\item \code{calc_Mf_1f()}: Calculate Williams M(f) ~ 1/f and saves it to
cevodata$models$Mf_1f

\item \code{calc_Mf_1f(cevodata)}: Method for \if{html}{\out{<cevodata>}} object

\item \code{calc_Mf_1f(cevo_snvs)}: Method for <cevo_snvs> object

\item \code{get_Mf_1f()}: Get M(f) ~ 1/f

Calculates M(f) ~ 1/f if not calculated yet

\item \code{plot_Mf_1f(cevodata)}: Plot M(f) ~ 1/f

}}
\examples{
data("test_data")
test_data |>
  calc_Mf_1f()

test_data |>
  plot_Mf_1f()
}
