% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_stats.R
\name{misc_stats}
\alias{misc_stats}
\alias{calc_all_sample_stats}
\alias{calc_all_patient_stats}
\title{Get extra sampe/patient stats}
\usage{
calc_all_sample_stats(object, snvs = default_SNVs(object))

calc_all_patient_stats(object, snvs = default_SNVs(object))
}
\arguments{
\item{object}{cevodata object}

\item{snvs}{Which SNVs to use}
}
\description{
Get extra sampe/patient stats
}
\section{Functions}{
\itemize{
\item \code{calc_all_sample_stats()}: Get misc sample level stats

\item \code{calc_all_patient_stats()}: Get misc patient level stats

}}
