% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quick_load}
\alias{quick_load}
\alias{quick_save}
\title{Quick save to ~/.cevomod directory}
\usage{
quick_load()

quick_save(object)
}
\arguments{
\item{object}{object to save}
}
\description{
Quick save to ~/.cevomod directory
}
\section{Functions}{
\itemize{
\item \code{quick_load()}: Quick load of ~/.cevomod/object.Rds

}}
