% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snvs.R
\name{set_snvs_frequency_measure}
\alias{set_snvs_frequency_measure}
\alias{get_snvs_frequency_measure}
\alias{set_snvs_frequency_measure.cevodata}
\alias{get_snvs_frequency_measure.cevodata}
\alias{set_snvs_frequency_measure.tbl_df}
\alias{get_snvs_frequency_measure.tbl_df}
\title{Set/Get SNVs frequency measure}
\usage{
set_snvs_frequency_measure(object, ...)

get_snvs_frequency_measure(object, ...)

\method{set_snvs_frequency_measure}{cevodata}(
  object,
  column_name,
  snvs_name = default_SNVs(object),
  ...
)

\method{get_snvs_frequency_measure}{cevodata}(object, snvs_name = default_SNVs(object), ...)

\method{set_snvs_frequency_measure}{tbl_df}(object, column_name, ...)

\method{get_snvs_frequency_measure}{tbl_df}(object, ...)
}
\arguments{
\item{object}{cevodata object}

\item{...}{Other arguments}

\item{column_name}{Name of new frequency measure column}

\item{snvs_name}{Name of SNVs assay}
}
\description{
Set/Get SNVs frequency measure
}
