% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevodata-class.R
\name{settings}
\alias{settings}
\alias{default_SNVs}
\alias{default_SNVs<-}
\alias{default_CNAs}
\alias{default_CNAs<-}
\alias{active_models}
\alias{active_models<-}
\title{Get/Set active assays of the cevodata object}
\usage{
default_SNVs(object, ...)

default_SNVs(object, ...) <- value

default_CNAs(object, ...)

default_CNAs(object, ...) <- value

active_models(object, ...)

active_models(object, ...) <- value
}
\arguments{
\item{object}{object}

\item{...}{other arguments}

\item{value}{name of new default assay}
}
\description{
Get/Set active assays of the cevodata object
}
\section{Functions}{
\itemize{
\item \code{default_SNVs()}: Get default SNVs assay of cevodata

\item \code{default_SNVs(object, ...) <- value}: Set default SNVs assay of cevodata

\item \code{default_CNAs()}: Get default CNAs assay of cevodata

\item \code{default_CNAs(object, ...) <- value}: Set default CNAs assay of cevodata

\item \code{active_models()}: Get active models from cevodata

\item \code{active_models(object, ...) <- value}: Set active models in cevodata

}}
