% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_SFS.R
\name{sfs}
\alias{sfs}
\alias{calc_SFS}
\alias{calc_SFS.cevodata}
\alias{calc_SFS.cevo_snvs}
\alias{get_SFS}
\alias{plot_SFS}
\alias{plot_SFS.cevodata}
\title{Site Frequency Spectra}
\usage{
calc_SFS(object, ...)

\method{calc_SFS}{cevodata}(
  object,
  name = "SFS",
  snvs_name = default_SNVs(object),
  column = get_snvs_frequency_measure(object, snvs_name),
  bins = NULL,
  verbose = get_verbosity(),
  ...
)

\method{calc_SFS}{cevo_snvs}(
  object,
  column = get_snvs_frequency_measure(object),
  bins = NULL,
  verbose = get_verbosity(),
  ...
)

get_SFS(object, name = "SFS", verbose = get_verbosity(), ...)

plot_SFS(object, ...)

\method{plot_SFS}{cevodata}(object, ..., mapping = NULL, name = "SFS", geom = "bar")
}
\arguments{
\item{object}{SNVs tibble object}

\item{...}{other arguments}

\item{column}{VAF or CCF/2}

\item{bins}{Resolution of the cumulative tails calculation}

\item{verbose}{verbose?}

\item{mapping}{aes()}

\item{geom}{Geom}

\item{which_snvs}{Which SNVs to use?}

\item{model_name}{name of slot with SFS statistics}
}
\description{
Site Frequency Spectra (or Variant Allele Frequency Spectra) are the main
statistic used by cevomod.
}
\section{Functions}{
\itemize{
\item \code{calc_SFS()}: Calculates spectra for all samples and saves and saves them
in cevodata$models$SFS tibble.

SFS columns description:
\itemize{
\item y number of mutations in the frequency interval
\item y_scaled with y values scaled to the range 0-1
}

\item \code{calc_SFS(cevodata)}: method for cevodata object

\item \code{calc_SFS(cevo_snvs)}: method for cevo_snvs object

\item \code{get_SFS()}: Get SFS

Calculates SFS if sfs slot in object$stats is empty

\item \code{plot_SFS()}: Plot SFS

\item \code{plot_SFS(cevodata)}: Plot SFS

}}
\examples{
data("test_data")

test_data |>
  calc_SFS() |>
  plot_SFS() +
  layer_mutations(test_data, mapping = ggplot2::aes(x = VAF), drivers = "BRCA")
}
