% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevodata-transformations.R, R/generics.R
\name{split_by.cevodata}
\alias{split_by.cevodata}
\alias{split_by}
\title{Split object}
\usage{
\method{split_by}{cevodata}(object, var, ...)

split_by(object, ...)
}
\arguments{
\item{object}{object to split}

\item{var}{name of column in metadata}

\item{...}{other arguments}
}
\description{
Split object
}
\section{Methods (by class)}{
\itemize{
\item \code{split_by(cevodata)}: Split cevodata object

This function is based on \code{filter.cevodata()}, thus might be inefficient
if there are too many levels in the var column.

}}
