% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snvs.R
\name{unite_mutation_id}
\alias{unite_mutation_id}
\title{Unite many columns to create mutation_id column}
\usage{
unite_mutation_id(snvs, sep = "-", include_gene_symbol = FALSE, remove = TRUE)
}
\arguments{
\item{snvs}{SNVs}

\item{sep}{Separator}

\item{include_gene_symbol}{Include gene symbol in mutation_id?}

\item{remove}{Remove united columns?}
}
\description{
Unite many columns to create mutation_id column
}
