% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevodata-class.R
\name{update_cevodata_v2_to_v3}
\alias{update_cevodata_v2_to_v3}
\title{Update to cevodata to v3}
\usage{
update_cevodata_v2_to_v3(cd)
}
\arguments{
\item{cd}{\if{html}{\out{<cevodata>}} object}
}
\description{
Update to cevodata to v3
}
