% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevodata-class.R, R/generics.R, R/snvs.R
\name{assays}
\alias{assays}
\alias{add_SNV_data}
\alias{add_CNA_data}
\alias{SNVs.cevodata}
\alias{CNAs.cevodata}
\alias{SNVs}
\alias{CNAs}
\alias{get_SNVs_wider}
\alias{get_SNVs_2d_matrix}
\title{Get/Add SNV/CNA data from the cevodata dataset}
\usage{
add_SNV_data(object, snvs, name = NULL)

add_CNA_data(object, cnas, name = NULL)

\method{SNVs}{cevodata}(object, which = default_SNVs(object), ...)

\method{CNAs}{cevodata}(object, which = default_CNAs(object), ...)

SNVs(object, ...)

CNAs(object, ...)

get_SNVs_wider(object, fill_na = NULL)

get_SNVs_2d_matrix(
  object,
  rows_sample = NULL,
  cols_sample = NULL,
  bins = NULL,
  verbose = verbose::verbose("cevoverse")
)
}
\arguments{
\item{object}{cevodata object}

\item{snvs}{tibble with SNVs}

\item{name}{name for SNVs/CNAs assay}

\item{cnas}{tibble with CNAs}

\item{which}{assay to use - uses active_SNVs if none}

\item{fill_na}{fill missing with this value}

\item{rows_sample}{Sample to use as rows}

\item{cols_sample}{Sample to use as cols}

\item{bins}{Number of bins to use for VAF intervals}

\item{verbose}{Verbose?}
}
\value{
tibble
}
\description{
Get/Add SNV/CNA data from the cevodata dataset
}
\section{Functions}{
\itemize{
\item \code{add_SNV_data()}: Add new SNVs to cevodata

\item \code{add_CNA_data()}: Add new CNAs to cevodata

\item \code{SNVs(cevodata)}: Get SNVs from cevodata dataset

\item \code{CNAs(cevodata)}: Get CNAs from cevodata dataset

\item \code{get_SNVs_wider()}: Get SNVs in the wide table form

\item \code{get_SNVs_2d_matrix()}: Get SNVs matrix (for single-patient cevodata only)

Requires 'sample' column in metadata.

}}
