% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevodata-package.R
\docType{package}
\name{cevodata-package}
\alias{cevodata}
\alias{cevodata-package}
\title{cevodata: Class for storing cancer DNAseq data}
\description{
cevodata provides a <cevodata> class, an R class for storing the DNA sequencing data for many patients and many tumor samples in a signle object, and methods to manipulate it.
}
\author{
\strong{Maintainer}: Paweł Kuś \email{kpawel2210@gmail.com} (\href{https://orcid.org/0000-0002-4367-9821}{ORCID})

}
\keyword{internal}
