% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_CNAs_var_names}
\alias{get_CNAs_var_names}
\alias{get_CNAs_var_names.cevodata}
\title{Get names of CNA variables}
\usage{
get_CNAs_var_names(object, ...)

\method{get_CNAs_var_names}{cevodata}(object, which = default_CNAs(object), ...)
}
\arguments{
\item{object}{object}

\item{...}{other arguments}

\item{which}{CNA assay to use}
}
\description{
Get names of CNA variables
}
\section{Methods (by class)}{
\itemize{
\item \code{get_CNAs_var_names(cevodata)}: Get CNA variable names from cevodata object

}}
