% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_stats.R
\name{misc_stats}
\alias{misc_stats}
\alias{calc_all_sample_stats}
\alias{calc_all_patient_stats}
\alias{calc_sample_sample_stats}
\title{Calc extra sample/patient stats}
\usage{
calc_all_sample_stats(object, snvs_name = default_SNVs(object))

calc_all_patient_stats(object, snvs_name = default_SNVs(object))

calc_sample_sample_stats(object, snvs_name = default_SNVs(object))
}
\arguments{
\item{object}{cevodata object}

\item{snvs_name}{Which SNVs to use}
}
\description{
Calculated stats are saved to appropriate cevodata$stats slots and can be
obtained with get_stats() function.
}
\section{Functions}{
\itemize{
\item \code{calc_all_sample_stats()}: Calc misc sample level stats

Currently calculates only the sample mutation burden

\item \code{calc_all_patient_stats()}: Calc misc patient level stats

Currently calculates only the patient mutation burden

\item \code{calc_sample_sample_stats()}: Calc sample-sample stats

Currently calculates only the Jaccard index

}}
