% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{split_by}
\alias{split_by}
\alias{split_by.cevodata}
\title{Split object}
\usage{
split_by(object, ...)

\method{split_by}{cevodata}(object, var, ...)
}
\arguments{
\item{object}{cevodata object}

\item{...}{other arguments}

\item{var}{name of column in metadata}
}
\description{
Split object
}
\section{Methods (by class)}{
\itemize{
\item \code{split_by(cevodata)}: Split cevodata object

This function is based on \code{filter.cevodata()}, thus might be inefficient
if there are too many levels in the var column.

}}
